fix(memory-safety): Phase 2 — frame bounds, UB-free shifts, BLE adv lifecycle, telnet fd
Closes audit-V3 findings C5, C6, C7, C8 plus a couple of bystanders
in the same files.

nivona_ble.c (C5):
  - nivona_ble_start_advertising now calls ble_gap_adv_stop()
    idempotently before adv_start. On a failed-connect restart
    (BLE_GAP_EVENT_CONNECT with status != 0) the controller can
    still be in adv-active state for a brief window and adv_start
    would return BLE_HS_EALREADY. We deliberately ignore EALREADY
    from adv_stop because both paths leave us in the desired
    "ready to start fresh" state.

nivona_frame.c (C6):
  - nivona_frame_send computes the required output size BEFORE any
    write into the static frame[] buffer. cmd length is enforced
    to [1,2] (any 3-byte opcode added in the future would have
    silently slid past every downstream check). On overflow we
    log and bail rather than corrupt the static buffer for the
    next caller.

nivona_dispatch.c (C7):
  - be16_i and handle_hw cast each uint8_t to uint32_t before the
    left shift, then back to int16_t/int32_t. The previous code
    relied on the C integer-promotion rules: req[2] is promoted to
    (signed) int, and on a 16-bit int platform a left shift by 8
    or 24 raises the sign bit, which is UB per C11 6.5.7. ESP32
    int is 32-bit so the immediate result was safe in practice,
    but the explicit cast is portable and silences UBSAN.

nivona_telnet.c (C8 + bystanders):
  - Client eviction now does shutdown(SHUT_RDWR) on the old fd
    so the old client_task breaks out of its blocking recv() with
    a clean EOF and proceeds to its own close(). Previously
    listen_task close()d the fd directly while the old task was
    still in recv(); lwIP could recycle the fd number before the
    old task observed the error, at which point the old task
    would read/write the new client's socket.
  - vsnprintf truncation: cap n at sizeof(buf)-1 (was sizeof(buf)),
    so the trailing NUL is not sent down the telnet stream.
  - socket() return value check on listen_task — without it
    setsockopt/bind both fail on fd -1 and the root cause is lost.

+++ main/nivona_dispatch.c
@@ -27,7 +27,13 @@ uint32_t g_diag_frame_parsed = 0;
 // ---- BE helpers --------------------------------------------------------
 
 static int16_t be16_i(const uint8_t *p) {
-    return (int16_t)((p[0] << 8) | p[1]);
+    // Cast each uint8_t to uint32_t before the shift. The C integer
+    // promotion otherwise raises p[0] to (signed) int, and on a 16-bit
+    // int platform that would left-shift into the sign bit — UB per
+    // C11 6.5.7. On ESP32 int is 32-bit so the immediate result is
+    // safe in practice, but the explicit cast makes the code portable
+    // and silences UBSAN.
+    return (int16_t)(((uint32_t)p[0] << 8) | (uint32_t)p[1]);
 }
 static void put_be16(uint8_t *p, int16_t v) {
     p[0] = (uint8_t)((v >> 8) & 0xFF); p[1] = (uint8_t)(v & 0xFF);
@@ -171,8 +177,14 @@ static void handle_hr(const uint8_t *req, size_t len) {
 static void handle_hw(const uint8_t *req, size_t len) {
     if (len >= 6) {
         int16_t id = be16_i(req);
-        int32_t v = (int32_t)((req[2] << 24) | (req[3] << 16) |
-                              (req[4] << 8)  |  req[5]);
+        // Cast each uint8_t to uint32_t before shifting — promoting
+        // req[2] >= 0x80 into the sign bit of a signed int is UB per
+        // C11 6.5.7. Cast the assembled value back to int32_t for the
+        // store API.
+        int32_t v = (int32_t)(((uint32_t)req[2] << 24) |
+                              ((uint32_t)req[3] << 16) |
+                              ((uint32_t)req[4] << 8)  |
+                               (uint32_t)req[5]);
         nivona_store_set_num(id, v);
         ESP_LOGI(TAG, "HW id=%d value=%ld", id, (long)v);
     }
